+++ teardrop/takecare/App.config
@@ -38,7 +38,7 @@
         <value />
       </setting>
       <setting name="extension" serializeAs="String">
-        <value>.fallentear</value>
+        <value>.toasty</value>
       </setting>
       <setting name="db_host" serializeAs="String">
         <value>127.0.0.1</value>
@@ -58,9 +58,6 @@
       <setting name="debug" serializeAs="String">
         <value>False</value>
       </setting>
-      <setting name="application_title" serializeAs="String">
-        <value>Your Title Here</value>
-      </setting>
       <setting name="disable_taskmgr" serializeAs="String">
         <value>False</value>
       </setting>
@@ -78,6 +75,9 @@
       </setting>
       <setting name="theme" serializeAs="String">
         <value>flash</value>
+      </setting>
+      <setting name="application_title" serializeAs="String">
+        <value />
       </setting>
     </teardrop.Properties.Settings>
   </userSettings>
