+++ teardrop/takecare/FrmMain.resx
