+++ teardrop/takecare/Properties/AssemblyInfo.cs
