+++ teardrop/takecare/Properties/Resources.resx
