+++ teardrop/takecare/app.manifest
