+++ teardrop/takecare/packages.config
