--- teardrop/teardrop/bin/Debug/message.html
@@ -1,1 +1,0 @@
-<h1>Github:</h1><p>hackthedev/teardrop</p>
