--- teardrop/teardrop/bin/Debug/teardrop.exe.config
@@ -1,84 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <configSections>
-        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="teardrop.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
-        </sectionGroup>
-    </configSections>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
-    </startup>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="System.Buffers" publicKeyToken="cc7b13ffcd2ddd51" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-4.0.3.0" newVersion="4.0.3.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Google.Protobuf" publicKeyToken="a7d26565bac4d604" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-3.15.8.0" newVersion="3.15.8.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="K4os.Compression.LZ4.Streams" publicKeyToken="2186fa9121ef231d" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-1.2.6.0" newVersion="1.2.6.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="BouncyCastle.Crypto" publicKeyToken="0e99375e54769942" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-1.8.9.0" newVersion="1.8.9.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="System.Runtime.CompilerServices.Unsafe" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-5.0.0.0" newVersion="5.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-  <userSettings>
-    <teardrop.Properties.Settings>
-      <setting name="key" serializeAs="String">
-        <value />
-      </setting>
-      <setting name="extension" serializeAs="String">
-        <value>.fallentear</value>
-      </setting>
-      <setting name="db_host" serializeAs="String">
-        <value>127.0.0.1</value>
-      </setting>
-      <setting name="db_user" serializeAs="String">
-        <value>root</value>
-      </setting>
-      <setting name="db_pass" serializeAs="String">
-        <value />
-      </setting>
-      <setting name="db_database" serializeAs="String">
-        <value>teardrop</value>
-      </setting>
-      <setting name="db_enable" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="debug" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="application_title" serializeAs="String">
-        <value>Your Title Here</value>
-      </setting>
-      <setting name="disable_taskmgr" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="clickMouse" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="message" serializeAs="String">
-        <value>&lt;h1&gt;Github:&lt;/h1&gt;&lt;p&gt;hackthedev/teardrop&lt;/p&gt;</value>
-      </setting>
-      <setting name="unkillable" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="application_name" serializeAs="String">
-        <value />
-      </setting>
-      <setting name="theme" serializeAs="String">
-        <value>flash</value>
-      </setting>
-    </teardrop.Properties.Settings>
-  </userSettings>
-</configuration>
